refactor: :recycle: move some types
