+++ packages/workspace-fusion/src/lib/types/event.ts
@@ -3,3 +3,7 @@ import { FusionMediator } from './fusionController';
 export type OnWorkspaceReadyEvent<TData extends Record<PropertyKey, unknown>> = {
 	api: FusionMediator<TData>;
 };
+
+export interface WorkspaceOnClick<TData> {
+	item: TData;
+}
