--- packages/workspace-fusion/src/lib/types/onClick.ts
@@ -1,3 +1,0 @@
-export interface WorkspaceOnClick<TData> {
-	item: TData;
-}
